requirements.txt
discord-py-interactions
python-dotenv
google-cloud-translate
six